Adicionando algumas validações

+++ src/service/viagens.js
@@ -166,8 +166,9 @@ exports.get_by_id = async (req, res, next) => {
             return res.status(HttpStatus.NOT_FOUND).json({
                 mensagem: 'Viagem não encontrada'
             })
-        }
-        res.status(HttpStatus.OK).json(convertViagem(viagem))
+        } else {
+            res.status(HttpStatus.OK).json(convertViagem(viagem))
+        }        
     })
     .catch(err => {
         res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
@@ -207,27 +208,45 @@ exports.iniciar = async (req, res, next) => {
     const { saida, km_inicial, descricao, veiculo, motorista } = req.body
 
     const veiculoBD = await Veiculo.findByPk(veiculo)
-    if (!veiculoBD) {
+
+    if (veiculo) {    
+        if (!veiculoBD)
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'Veículo inexistente'
+            })
+
+        if (!veiculoBD.dataValues.disponivel)
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'Veículo indisponível'
+            })
+    } else {
         return res.status(HttpStatus.BAD_REQUEST).json({
-            mensagem: 'Veículo inexistente'
-        })
-    }
-    if (!veiculoBD.dataValues.disponivel) {
-        return res.status(HttpStatus.BAD_REQUEST).json({
-            mensagem: 'Veículo indisponível'
+            mensagem: 'O veículo é obrigatório'
         })
     }
 
     const motoristaBD = await Motorista.findByPk(motorista)
-    if (!motoristaBD) {
+
+    if (motorista) {
+        if (!motoristaBD)
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'Motorista inexistente'
+            })
+
+        if (!motoristaBD.dataValues.disponivel)
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'Motorista indisponível'
+            })
+    } else {
         return res.status(HttpStatus.BAD_REQUEST).json({
-            mensagem: 'Motorista inexistente'
+            mensagem: 'O motorista é obrigatório'
         })
     }
-    if (!motoristaBD.dataValues.disponivel) {
-        return res.status(HttpStatus.BAD_REQUEST).json({
-            mensagem: 'Motorista indisponível'
-        })
+
+    const salvar = {
+        saida, km_inicial, descricao, 
+        id_veiculo: veiculo,
+        id_motorista: motorista
     }
 
     if (!checkCNH(motoristaBD.dataValues.categoria, veiculoBD.dataValues.cnh_requerida)) {
@@ -236,13 +255,37 @@ exports.iniciar = async (req, res, next) => {
         })
     }
 
-    Viagem.create({
-        saida,
-        km_inicial,
-        descricao,
-        id_veiculo: veiculo,
-        id_motorista: motorista
-    })
+    if (saida) {
+        const data = saida.replace("T", " ")
+        const padrao = /^[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}$/
+        const s = new Date(data)
+
+        if (!padrao.test(data) || isNaN(s.getTime())) {
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'A data não segue o padrão: yyyy-MM-dd HH:mm:ss'
+            })
+        }
+        salvar.saida = data
+    } else {
+        return res.status(HttpStatus.BAD_REQUEST).json({
+            mensagem: 'O momento da saída é obrigatório'
+        })
+    }
+
+    if (km_inicial) {
+        if (isNaN(km_inicial)) {
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'O valor da quilometragem é inválido'
+            })
+        }
+        salvar.km_inicial = parseFloat(km_inicial)
+    } else {
+        return res.status(HttpStatus.BAD_REQUEST).json({
+            mensagem: 'O KM inicial é obrigatório'
+        })
+    }
+
+    Viagem.create(salvar)
     .then(viagem => {
         Veiculo.update({
             disponivel: false
@@ -252,15 +295,17 @@ exports.iniciar = async (req, res, next) => {
             disponivel: false
         }, { where: { id: motorista } })
 
-        const result = {
-            id: viagem.id,
-            saida: viagem.saida,
-            km_inicial: viagem.km_inicial,
-            descricao: viagem.descricao,
-            veiculo: viagem.id_veiculo,
-            motorista: viagem.id_motorista
-        }
-        return res.status(HttpStatus.CREATED).json(result)
+        Viagem.findByPk(viagem.id, {
+            include: [Veiculo, Motorista] 
+        })
+        .then(viagem => {
+            res.status(HttpStatus.OK).json(convertViagem(viagem))
+        })
+        .catch(err => {
+            res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
+                mensagem: err
+            })
+        })
     })
     .catch(err => {
         return res.status(HttpStatus.BAD_REQUEST).json({
@@ -271,22 +316,92 @@ exports.iniciar = async (req, res, next) => {
 
 exports.concluir = async (req, res, next) => {
     const { viagemId } = req.params
-    const { saida, chegada, km_inicial, km_final, descricao, veiculo, motorista } = req.body
-
+    
     const viagem = await Viagem.findByPk(viagemId)
     if (!viagem) {
         return res.status(HttpStatus.BAD_REQUEST).json({
             mensagem: 'Viagem não encontrada'
         })
     }  
-    console.log(`KM FINAL: ${km_final}, KM INICIAL: ${km_inicial}`)
-    if (parseFloat(km_final) < parseFloat(km_inicial)) {
+
+    const { saida, chegada, km_inicial, km_final, descricao } = req.body
+    const salvar = {
+        saida, chegada, km_final, km_inicial, descricao
+    }
+
+    const padrao = /^[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}$/
+
+    if (saida) {
+        const aux = saida.replace("T", " ")    
+        const s = new Date(aux)
+
+        if (!padrao.test(aux) || isNaN(s.getTime())) {
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'A data de saída não segue o padrão: yyyy-MM-dd HH:mm:ss'
+            })
+        }
+    } else {
+        delete salvar.saida
+    }
+
+    if (km_inicial) {
+        if (isNaN(km_inicial)) {
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'O valor da quilometragem inicial é inválido'
+            })
+        }
+        salvar.km_inicial = parseFloat(km_inicial)
+    } else {
+        delete salvar.km_inicial
+    }
+
+    if (km_final) {
+        if (isNaN(km_final)) {
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'O valor da quilometragem final é inválido'
+            })
+        }
+        salvar.km_final = parseFloat(km_final)
+    } else {
         return res.status(HttpStatus.BAD_REQUEST).json({
-            mensagem: `KM final (${km_final}Km) não pode ser menor que KM inicial (${km_inicial}Km)`
+            mensagem: 'A quilometragem final é obrigatória'
         })
     }
 
-    const dataSaida = new Date(saida)
+    var inicialKm = km_inicial
+    if (!inicialKm) {
+        inicialKm = viagem.dataValues.km_inicial
+    }
+
+    console.log(`KM FINAL: ${km_final}, KM INICIAL: ${inicialKm}`)
+    if (parseFloat(km_final) < parseFloat(inicialKm)) {
+        return res.status(HttpStatus.BAD_REQUEST).json({
+            mensagem: `KM final (${km_final}KM) não pode ser menor que KM inicial (${inicialKm}KM)`
+        })
+    }
+
+    if (chegada) {
+        const aux = chegada.replace("T", " ")
+        const s = new Date(aux)
+
+        if (!padrao.test(aux) || isNaN(s.getTime())) {
+            return res.status(HttpStatus.BAD_REQUEST).json({
+                mensagem: 'A data de chegada não segue o padrão: yyyy-MM-dd HH:mm:ss'
+            })
+        }
+    } else {
+        return res.status(HttpStatus.BAD_REQUEST).json({
+            mensagem: 'A momento de chegada é obrigatória'
+        })
+    }
+
+    var saidaAux = saida
+
+    if (!saidaAux) {
+        saidaAux = viagem.dataValues.saida // Se ele não enviou agora eu pego a data salva em banco
+    }
+
+    const dataSaida = new Date(saidaAux)
     const dataChegada = new Date(chegada)
 
     if (dataChegada < dataSaida) {
@@ -295,51 +410,70 @@ exports.concluir = async (req, res, next) => {
         })
     }
 
-    // Validar informações
+    Viagem.update(salvar, 
+        { where: { id: viagemId } })
+    .then(() => {
 
-    Viagem.update({
-        saida, // verificar se vai poder atualizar
-        chegada, 
-        km_inicial, // verificar se vai poder atualizar
-        km_final,
-        descricao,
-        // id_veiculo: veiculo,
-        // id_motorista: motorista
-    }, { where: { id: viagemId } })
-    .then(viagem => {
-        Veiculo.update({
-            disponivel: true,
-            quilometragem: km_final
-        }, { where: { id: veiculo } })
-
-        Motorista.update({
-            disponivel: true
-        }, { where: { id: motorista } })
-
-        Viagem.findByPk(viagemId)
+        Viagem.findByPk(viagemId, {
+            include: [Veiculo, Motorista]
+        })
         .then(viagem => {
-            return res.status(HttpStatus.OK).json(viagem)
+
+            Veiculo.update({
+                disponivel: true,
+                quilometragem: salvar.km_final
+            }, { where: { id: viagem.dataValues.id_veiculo } })
+
+            Motorista.update({
+                disponivel: true
+            }, { where: { id: viagem.dataValues.id_motorista } })
+
+            return res.status(HttpStatus.OK).json(convertViagem(viagem))
         })
         .catch(err => {
+            console.log(err)
             return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 mensagem: err
             })
         })        
     })
     .catch(err => {
+        console.log(err)
         return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
             mensagem: err
         })
     })
 }
 
-exports.deletar = (req, res, next) => {
+exports.deletar = async (req, res, next) => {
     const id = req.params.viagemId
+
+    const viagem = await Viagem.findByPk(id)
+    if (!viagem) {
+        return res.status(HttpStatus.NOT_FOUND).json({
+            mensagem: 'Viagem não encontrada'
+        })
+    }
+
+    const veiculoId = viagem.dataValues.id_veiculo
+    const motoristaId = viagem.dataValues.id_motorista
+    const nao_concluida = viagem.dataValues.chegada == null
 
     Viagem.destroy({
         where: { id }
     })
-    .then(() => {
+    .then(rows => {
+        if (rows > 0) {
+            if (nao_concluida) {
+                Veiculo.update({
+                    disponivel: true
+                }, { where: { id: veiculoId } })
+    
+                Motorista.update({
+                    disponivel: true
+                }, { where: { id: motoristaId } })
+            }
+        }
         res.status(HttpStatus.NO_CONTENT).send()
     })
     .catch(err => {
